+++ ui/src/Hooks/useFetchPersonTags/useFetchPersonTags.tsx
@@ -18,7 +18,7 @@
 import {useRecoilState} from 'recoil';
 import {useCallback} from 'react';
 import sortTagsAlphabetically from 'Tags/sortTagsAlphabetically';
-import {RoleTag, Tag} from 'Types/Tags';
+import {RoleTag, Tag} from 'Types/Tag';
 import {PersonTagsState} from 'State/PersonTagsState';
 import PersonTagClient from 'Tags/PersonTag/PersonTagClient';
