+++ ui/src/ModalFormComponents/ViewTagRow.tsx
@@ -19,7 +19,7 @@ import {JSX} from '@babel/types';
 import React, {ReactNode} from 'react';
 import {createDataTestId} from '../Utils/ReactUtils';
 import {TagType} from '../SortingAndFiltering/FilterLibraries';
-import {Tag} from 'Types/Tags';
+import {Tag} from 'Types/Tag';
 
 interface Props {
     children?: ReactNode;
