+++ ui/src/Products/ProductForm.tsx
@@ -21,7 +21,7 @@ import moment from 'moment';
 import ProductClient from './ProductClient';
 import {emptyProduct} from './ProductService';
 import ConfirmationModal, {ConfirmationModalProps} from 'Modal/ConfirmationModal/ConfirmationModal';
-import {Tag} from 'Types/Tags';
+import {Tag} from 'Types/Tag';
 import ProductFormLocationField from './ProductFormLocationField';
 import ProductFormStartDateField from './ProductFormStartDateField';
 import ProductFormEndDateField from './ProductFormEndDateField';
